+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.6.0 - ###-##-##
 - Upgrade internal API to v2.0.
 - Add experimental methods AMPL._startRecording and AMPL._stopRecording.
+- Add experimental method AMPL._loadSession.
 
 ## 0.5.0 - 2018-05-06
 - Breaking: AMPL errors raise exceptions by default.